+++ src/Translation/Writer.php
@@ -7,6 +7,7 @@ use Illuminate\Foundation\Application;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Nikaia\TranslationSheet\Commands\Output;
+use Nikaia\TranslationSheet\Sheet\TranslationsSheet;
 use Nikaia\TranslationSheet\Spreadsheet;
 use Nikaia\TranslationSheet\Util;
 
@@ -26,6 +27,9 @@ class Writer
     /** @var Application */
     protected $app;
 
+    /** @var TranslationsSheet */
+    protected $translationSheet;
+
     public function __construct(Spreadsheet $spreadsheet, Filesystem $files, Application $app)
     {
         $this->spreadsheet = $spreadsheet;
@@ -34,6 +38,20 @@ class Writer
 
         $this->nullOutput();
     }
+
+    /**
+     * Set the translation sheets for reader.
+     *
+     * @param TranslationsSheet $translationSheet
+     * @return Writer
+     */
+    public function setTranslationsSheet(TranslationsSheet $translationSheet)
+    {
+        $this->translationSheet = $translationSheet;
+
+        return $this;
+    }
+
 
     public function setTranslations($translations)
     {
@@ -46,7 +64,16 @@ class Writer
     {
         $this
             ->groupTranslationsByFile()
-            ->map(function ($items, $sourceFile) {
+            ->each(function ($items, $sourceFile) {
+
+                if ($this->files->extension($sourceFile) === 'json') {
+                    $this->writeJsonFile(
+                        $this->translationSheet->getPath() . '/' . $sourceFile,
+                        $items
+                    );
+                    return;
+                }
+
                 $this->writeFile(
                     $sourceFile,
                     $items
@@ -56,15 +83,25 @@ class Writer
 
     protected function writeFile($file, $items)
     {
-        $this->output->writeln('  '.$file);
+        $this->output->writeln('    <comment>php</comment> ' . $file);
 
-        $content = "<?php\n\nreturn ".Util::varExport($items).";\n";
+        $content = sprintf('<?php%s%sreturn %s;%s', PHP_EOL, PHP_EOL, Util::varExport($items), PHP_EOL);
 
-        if (! $this->files->isDirectory($dir = dirname($file))) {
+        if (!$this->files->isDirectory($dir = dirname($file))) {
             $this->files->makeDirectory($dir, 0755, true);
         }
 
         $this->files->put($file, $content);
+    }
+
+    protected function writeJsonFile($file, $items)
+    {
+        $this->output->writeln('    <comment>json</comment> ' . $file);
+
+        if (!$this->files->isDirectory($dir = dirname($file))) {
+            $this->files->makeDirectory($dir, 0755, true);
+        }
+        $this->files->put($file, json_encode($items, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES) . PHP_EOL);
     }
 
     protected function groupTranslationsByFile()
@@ -97,19 +134,41 @@ class Writer
                 // For instance, we have `app.title` that is the same for each locale,
                 // We dont want to translate it to every locale, and prefer letting
                 // Laravel default back to the default locale.
-                if (! isset($translation[$locale])) {
+                if ($this->skipToDefault($translation, $locale)) {
                     continue;
                 }
 
-                $localeFile = str_replace('{locale}/', $locale.'/', $translation['sourceFile']);
+                $localeFile = $this->fileIsJson($translation['sourceFile']) ?
+                    $locale . '.json' :
+                    str_replace('{locale}/', $locale . '/', $translation['sourceFile']);
+
                 if (empty($files[$localeFile])) {
                     $files[$localeFile] = [];
                 }
 
-                Arr::set($files[$localeFile], $translation['key'], $translation[$locale]);
+                if ($this->fileIsJson($translation['sourceFile'])) {
+                    $files[$localeFile][$translation['key']] = $translation[$locale];
+                } else {
+                    Arr::set($files[$localeFile], $translation['key'], $translation[$locale]);
+                }
             }
         }
 
         return $files;
     }
+
+    protected function skipToDefault($translation, $locale)
+    {
+        if (!isset($translation[$locale])) {
+            return true;
+        }
+
+        return empty($translation[$locale]) && $this->fileIsJson($translation['sourceFile']);
+    }
+
+    protected function fileIsJson($sourceFile)
+    {
+        return $sourceFile === '{locale}.json';
+    }
+
 }