+++ tests/Feature/ExcludePatternsTest.php
@@ -1,9 +1,16 @@
 <?php
 
-use Nikaia\TranslationSheet\Pusher;
+namespace Nikaia\TranslationSheet\Test\Feature;
+
+use Nikaia\TranslationSheet\Commands\Prepare;
+use Nikaia\TranslationSheet\Commands\Push;
+use Nikaia\TranslationSheet\Commands\Setup;
+use Nikaia\TranslationSheet\SheetPusher;
+use Nikaia\TranslationSheet\Sheet\TranslationsSheet;
+use Nikaia\TranslationSheet\Spreadsheet;
 use Nikaia\TranslationSheet\Test\FeatureTestCase;
 
-class ExcludeFilterTest extends FeatureTestCase
+class ExcludePatternsTest extends FeatureTestCase
 {
     protected function getEnvironmentSetUp($app)
     {
@@ -13,21 +20,49 @@ class ExcludeFilterTest extends FeatureTestCase
     }
 
     /** @test */
-    public function it_exludes_correctly_the_specified_patterns()
+    public function it_excludes_correctly_the_specified_patterns()
     {
-        /** @var Pusher $pusher */
-        $pusher = resolve(Pusher::class);
+        $this->helper->noExtraTranslationSheet();
+
+        /** @var SheetPusher $pusher */
+        $pusher = resolve(SheetPusher::class)->setTranslationsSheet(
+            Spreadsheet::primaryTranslationSheet()
+        );
 
         config()->set('translation_sheet.exclude', [
             'foo::*',
-            'validation*'
+            'validation*',
         ]);
         $this->assertCount(0, $pusher->getScannedAndTransformedTranslations());
-
 
         config()->set('translation_sheet.exclude', [
             'foo::*',
         ]);
         $this->assertCount(4, $pusher->getScannedAndTransformedTranslations());
+    }
+
+    /** @test */
+    public function it_excludes_correctly_the_specified_patterns_for_push()
+    {
+        $this->helper->noExtraTranslationSheet();
+
+        /** @var SheetPusher $pusher */
+        $pusher = resolve(SheetPusher::class)->setTranslationsSheet(
+            Spreadsheet::primaryTranslationSheet()
+        );
+        config()->set('translation_sheet.exclude', [
+            'foo::*',
+        ]);
+        $this->assertIsArray(json_decode($pusher->getScannedAndTransformedTranslations()->toJson()));
+
+        $this->resetSpreadsheet();
+
+        foreach ([
+                     (new Setup),
+                     (new Prepare),
+                     (new Push)
+                 ] as $command) {
+            $this->artisan($command->getName())->assertExitCode(0);
+        }
     }
 }